+++ src/constraints/indicator.jl
@@ -15,11 +15,15 @@ function init_constraint!(
     set::IS
 ) where {A, T<:Real, ASS<:MOI.AbstractScalarSet, IS<:Union{IndicatorSet{A}, MOI.IndicatorSet{A, ASS}}}
     inner_constraint = constraint.inner_constraint
-    if hasmethod(	
-        init_constraint!,	
-        (CS.CoM, typeof(inner_constraint), typeof(inner_constraint.std.fct), typeof(inner_constraint.std.set)),	
-    )	
-        return init_constraint!(com, inner_constraint, inner_constraint.std.fct, inner_constraint.std.set)
+
+    # check which methods that inner constraint supports
+    set_impl_functions!(com, inner_constraint)
+
+    if inner_constraint.std.impl.init
+        feasible = init_constraint!(com, inner_constraint, inner_constraint.std.fct, inner_constraint.std.set; active=false)
+        # map the bounds to the indicator constraint
+        constraint.std.bound_rhs = inner_constraint.std.bound_rhs
+        return feasible
     end
     # still feasible
     return true
@@ -114,4 +118,102 @@ function is_solved_constraint(
         return is_solved_constraint(inner_constraint, inner_constraint.std.fct, inner_constraint.std.set, values[2:end])
     end
     return true
+end
+
+
+"""
+    update_best_bound_constraint!(com::CS.CoM,
+        constraint::IndicatorConstraint,
+        fct::Union{MOI.VectorOfVariables, VAF{T}},
+        set::IS,
+        var_idx::Int,
+        lb::Int,
+        ub::Int
+    ) where {A, T<:Real, ASS<:MOI.AbstractScalarSet, IS<:Union{IndicatorSet{A}, MOI.IndicatorSet{A, ASS}}}
+
+Update the bound constraint associated with this constraint. This means that the `bound_rhs` bounds will be changed according to 
+the possible values the table constraint allows. `var_idx`, `lb` and `ub` don't are not considered atm.
+Additionally only a rough estimated bound is used which can be computed relatively fast. 
+This method calls the inner_constraint method if it exists and the indicator is activated.
+"""
+#=
+function update_best_bound_constraint!(com::CS.CoM,
+    constraint::IndicatorConstraint,
+    fct::Union{MOI.VectorOfVariables, VAF{T}},
+    set::IS,
+    var_idx::Int,
+    lb::Int,
+    ub::Int
+) where {A, T<:Real, ASS<:MOI.AbstractScalarSet, IS<:Union{IndicatorSet{A}, MOI.IndicatorSet{A, ASS}}}
+    inner_constraint = constraint.inner_constraint
+    indicator_var_idx = constraint.std.indices[1]
+    search_space = com.search_space
+    indicator_var = search_space[indicator_var_idx]
+    if inner_constraint.std.impl.update_best_bound
+        if CS.issetto(indicator_var, Int(constraint.activate_on)) 
+            return update_best_bound_constraint!(com, inner_constraint, inner_constraint.std.fct, inner_constraint.std.set, var_idx, lb, ub)
+        else
+            # if not activated (for example in a different subtree) we reset the bounds
+            for rhs in constraint.std.bound_rhs
+                rhs.lb = typemin(Int64)
+                rhs.ub = typemax(Int64)
+            end
+        end
+    end
+    return true
+end
+=#
+
+function single_reverse_pruning_constraint!(
+    com::CoM,
+    constraint::IndicatorConstraint,
+    fct::Union{MOI.VectorOfVariables, VAF{T}},
+    set::IS,
+    var::Variable,
+    backtrack_idx::Int
+) where {A, T<:Real, ASS<:MOI.AbstractScalarSet, IS<:Union{IndicatorSet{A}, MOI.IndicatorSet{A, ASS}}}
+    inner_constraint = constraint.inner_constraint
+    # the variable must be part of the inner constraint
+    # General assumption is that the indicator constraint is not part of the inner constraint
+    # TODO: Maybe disallow that indicator is part of inner constraint
+    if inner_constraint.std.impl.single_reverse_pruning && var.idx != constraint.std.indices[1]
+        single_reverse_pruning_constraint!(com, inner_constraint, inner_constraint.std.fct, inner_constraint.std.set, var, backtrack_idx)
+    end
+end
+
+function reverse_pruning_constraint!(
+    com::CoM,
+    constraint::IndicatorConstraint,
+    fct::Union{MOI.VectorOfVariables, VAF{T}},
+    set::IS,
+    backtrack_id::Int
+) where {A, T<:Real, ASS<:MOI.AbstractScalarSet, IS<:Union{IndicatorSet{A}, MOI.IndicatorSet{A, ASS}}}
+    inner_constraint = constraint.inner_constraint
+    if inner_constraint.std.impl.reverse_pruning
+        reverse_pruning_constraint!(com, inner_constraint, inner_constraint.std.fct, inner_constraint.std.set, backtrack_id)
+    end
+end
+  
+function restore_pruning_constraint!(
+    com::CoM,
+    constraint::IndicatorConstraint,
+    fct::Union{MOI.VectorOfVariables, VAF{T}},
+    set::IS,
+    prune_steps::Union{Int, Vector{Int}}
+) where {A, T<:Real, ASS<:MOI.AbstractScalarSet, IS<:Union{IndicatorSet{A}, MOI.IndicatorSet{A, ASS}}}
+    inner_constraint = constraint.inner_constraint
+    if inner_constraint.std.impl.restore_pruning
+        restore_pruning_constraint!(com, inner_constraint, inner_constraint.std.fct, inner_constraint.std.set, prune_steps)
+    end
+end
+
+function finished_pruning_constraint!(com::CS.CoM,
+    constraint::IndicatorConstraint,
+    fct::Union{MOI.VectorOfVariables, VAF{T}},
+    set::IS
+) where {A, T<:Real, ASS<:MOI.AbstractScalarSet, IS<:Union{IndicatorSet{A}, MOI.IndicatorSet{A, ASS}}}
+    inner_constraint = constraint.inner_constraint
+    if inner_constraint.std.impl.finished_pruning
+        finished_pruning_constraint!(com, inner_constraint, inner_constraint.std.fct, inner_constraint.std.set)
+    end
 end